fix authorization header issue in apollo client
